still needs to only accept pokemon names
everything works totally well, but it accepts all full rows instead of just pokemon names.  I also want to add in a type box on the right hand side so that if you guess a pokemon with a type that's the same as in the answer it will give you the type.